+++ apps/qna/serializers/answer/common.py
@@ -1,13 +1,41 @@
 from rest_framework import serializers
 
-from apps.qna.utils.model_types import User
+from apps.qna.models import Answer, AnswerComment
+from apps.qna.serializers.common import AuthorSerializer
 
 
-class AnswerAuthorSerializer(serializers.Serializer[User]):
+# ==============================================================================
+# Answer Serializers for Response
+# ==============================================================================
+class AnswerCommentSerializer(serializers.ModelSerializer[AnswerComment]):
     """
-    답변 및 댓글 작성자 정보 시리얼라이저
+    답변 댓글 상세 정보 시리얼라이저
     """
 
-    id = serializers.IntegerField(help_text="작성자 ID")
-    nickname = serializers.CharField(help_text="작성자 닉네임")
-    profile_image_url = serializers.ImageField(source="profile_img_url", use_url=True, help_text="프로필 이미지 URL")
+    author = AuthorSerializer(read_only=True)
+    created_at = serializers.DateTimeField(format="%Y-%m-%d %H:%M:%S")
+
+    class Meta:
+        model = AnswerComment
+        fields = ["id", "content", "created_at", "author"]
+
+
+class AnswerSerializer(serializers.ModelSerializer[Answer]):
+    """
+    질문 상세 조회 내 답변 목록용 시리얼라이저
+    """
+
+    author = AuthorSerializer(read_only=True)
+    comments = AnswerCommentSerializer(many=True, read_only=True)
+    created_at = serializers.DateTimeField(format="%Y-%m-%d %H:%M:%S")
+
+    class Meta:
+        model = Answer
+        fields = [
+            "id",
+            "content",
+            "created_at",
+            "is_adopted",
+            "author",
+            "comments",
+        ]
